requirements.txt
requests==2.28.1
openpyxl==3.0.10
unittestreport==1.0.0
pycryptodome==3.15.0